ml-api/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
transformers==4.36.0
torch>=2.0.0
scikit-learn
pandas
numpy
redis==5.0.1
sentence-transformers==2.2.2
faiss-cpu
